// src/modules/clan/channel/enums/CustomIds.enum.ts
export enum ClanChannelButtonCustomIds {
    VoiceChannel = 'KSLpvh-8wAkqS-85bNBA-g9UkWA',
    TextChannel = '75XQFt-2hZFGs-PBvNM5-dMpSmH',
    UserRights = 'Mc9zuv-PHTCbK-dvw6bf-EW8wQ7',
    Name = 'm8aURB-Vjtw28-6zLXXa-zKweUJ',
    Delete = '992ipT-Fpr3Tm-uQGxgD-XXrmtY',

    LimitMembers = 'HCxpm3-P7w94J-WGWRLE-hmmsdn',
    Rights = '6MRHha-QhiS8T-HhwRiA-Qge4P5',
    KickUser = '9uAusf-ZM4HXv-3XgDsN-RXkYa6',
    Skip1 = 'Qm52`x-CU7h7n-BwMRCD-bUrQ,j',
    Skip2 = 'M4P:Dq-we9Jvq-8s8sqK-Rf(AFE',
}