// src/modules/settings/clan/channel/enums/ClanChannelEmoji.enum.ts
export enum ClanChannelEmoji {
    VoiceChannel = '<:create_voice:1221128605027926127>',
    TextChannel = '<:create_text:1221128626007838730>',
    UserRights = '<:user_rights:1221128615735853207>',
    Name = '<:edit_name:1221128621956141236>',
    Delete = '<:delete:1221128620022567025>',
    LimitMembers = '<:limit:1221128617640067252>',
    Rights = '<:add_rights:1221128612976005221>',
    KickUser = '<:user_kick:1221128610782511246>',
}